Moving files to the correct place and adding styles.
